refactor(cli):move client.go into client/cli and split as tx.go/query.go
