+++ x/sponsor-contract-tx/client/cli/tx.go
@@ -1,7 +1,6 @@
-package sponsor
+package cli
 
 import (
-	"context"
 	"fmt"
 	"math/big"
 	"regexp"
@@ -15,6 +14,26 @@ import (
 
 	"github.com/DoraFactory/doravota/x/sponsor-contract-tx/types"
 )
+
+
+// GetTxCmd returns the transaction commands for the sponsor module
+func GetTxCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:                        types.ModuleName,
+		Short:                      "Sponsor transaction subcommands",
+		DisableFlagParsing:         true,
+		SuggestionsMinimumDistance: 2,
+		RunE:                       client.ValidateCmd,
+	}
+
+	cmd.AddCommand(
+		GetCmdSetSponsor(),
+		GetCmdUpdateSponsor(),
+		GetCmdDeleteSponsor(),
+	)
+
+	return cmd
+}
 
 // parseCoinsWithDORASupport parses coins string with support for DORA to peaka conversion
 // 1 DORA = 10^18 peaka
@@ -64,44 +83,7 @@ func parseCoinsWithDORASupport(coinsStr string) (sdk.Coins, error) {
 	return coins, nil
 }
 
-// GetTxCmd returns the transaction commands for the sponsor module
-func GetTxCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:                        types.ModuleName,
-		Short:                      "Sponsor transaction subcommands",
-		DisableFlagParsing:         true,
-		SuggestionsMinimumDistance: 2,
-		RunE:                       client.ValidateCmd,
-	}
 
-	cmd.AddCommand(
-		GetCmdSetSponsor(),
-		GetCmdUpdateSponsor(),
-		GetCmdDeleteSponsor(),
-	)
-
-	return cmd
-}
-
-// GetQueryCmd returns the query commands for the sponsor module
-func GetQueryCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:                        types.ModuleName,
-		Short:                      "Sponsor query subcommands",
-		DisableFlagParsing:         true,
-		SuggestionsMinimumDistance: 2,
-		RunE:                       client.ValidateCmd,
-	}
-
-	cmd.AddCommand(
-		GetCmdQueryAllSponsors(),
-		GetCmdQuerySponsorStatus(),
-		GetCmdQueryUserGrantUsage(),
-		GetCmdQueryParams(),
-	)
-
-	return cmd
-}
 
 // GetCmdSetSponsor implements the set sponsor command
 func GetCmdSetSponsor() *cobra.Command {
@@ -194,124 +176,5 @@ func GetCmdDeleteSponsor() *cobra.Command {
 	}
 
 	flags.AddTxFlagsToCmd(cmd)
-	return cmd
-}
-
-// GetCmdQueryAllSponsors implements the query all-sponsors command
-func GetCmdQueryAllSponsors() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "all-sponsors",
-		Short: "Query all sponsor contracts",
-		Args:  cobra.NoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			req := &types.QueryAllSponsorsRequest{}
-			res, err := queryClient.AllSponsors(context.Background(), req)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-	return cmd
-}
-
-// GetCmdQuerySponsorStatus implements the query sponsor status command
-func GetCmdQuerySponsorStatus() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "status [contract-address]",
-		Short: "Query the status of a sponsor contract",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			req := &types.QuerySponsorRequest{
-				ContractAddress: args[0],
-			}
-			res, err := queryClient.Sponsor(context.Background(), req)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-	return cmd
-}
-
-// GetCmdQueryParams implements the query params command
-func GetCmdQueryParams() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "params",
-		Short: "Query the parameters of the sponsor module",
-		Args:  cobra.NoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			req := &types.QueryParamsRequest{}
-			res, err := queryClient.Params(context.Background(), req)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-	return cmd
-}
-
-// GetCmdQueryUserGrantUsage implements the query user grant usage command
-func GetCmdQueryUserGrantUsage() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "grant-usage [user-address] [contract-address]",
-		Short: "Query grant usage for a specific user and contract",
-		Long: `Query the grant usage information for a specific user and contract.
-This shows how much of the sponsor's grant the user has already consumed.`,
-		Args: cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			req := &types.QueryUserGrantUsageRequest{
-				UserAddress:     args[0],
-				ContractAddress: args[1],
-			}
-			res, err := queryClient.UserGrantUsage(context.Background(), req)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
 	return cmd
 }
